Add a check for incoming JWT in the authTokens array to auth middleware.
It occured to me while working on the login routes that I was only
checking that a auth token was valid, but not that it was actually
included in the users authTokens array. Once a user logs out, if
there is a JWT floating around somewhere for the user that is valid
but is not in the authTokens array, I don't want to allow access.