Remove GET /login route-handler and rewrite as POST /login.
I wasn't thinking correctly about the login routes. GET /login
would be a route that returns the login page if the user isn't
logged in, or redirects the user to the page they came from if
they are logged in. The correct route for logging a user in is
POST /login, which would be triggered on the login form submission.